Merge branch 'master' of https://github.com/prowill/CoordinateProcessor
Added a 2d Histogram function that generates a neat histogram from two input CSV files.

Still need to refactor main so that it doesnt generate histograms for all files, but first middle and last frames.

I have also been able to add some functionality to boil down the many csv files that are generated to an xmaster, zmaster and ymaster file. These files will be used in the next commit hopefully to generate a nice plot for Dr. Lindberg.

Signed-off-by: Prowill <william.robinson223@gmail.com>